requirements.txt
prefect~=0.15.9
requests~=2.26.0
pandas~=2.0.3
python-dateutil~=2.8.2
SQLAlchemy~=2.0.19
pytz~=2023.3
